feat: share scene recommendation pipeline across web and skill
