+++ README.md
@@ -132,6 +132,7 @@ python scripts/run_webui.py --reload
 - 人物校对页，支持关键字段补全、证据不足检查与二级字段微调
 - 场景卡、角色卡、开局模板的创建、编辑、选择与复用
 - 聊天过程中自动推荐下一幕场景卡，并支持会话内切换场景
+- 拍点成熟时自动浮出下一幕建议，附带转场起句、后续戏路和自动起拍提示
 - 会话恢复、最近会话续聊、群聊继续与工作台直接入场
 - 对话上下文自动压缩，按活跃角色裁剪人物/关系上下文，并注入会话记忆摘要
 - 在同一页面查看 transcript、继续群聊、删除历史会话
@@ -217,6 +218,7 @@ python scripts/run_webui.py --reload
 - 角色卡：为 `insert` 模式准备你的身份、语气、动机与在场定位
 - 开局模板：把入场方式、参与角色、场景卡、角色卡打包成一套可复用开局
 - 自动场景推荐：会话进行中，系统会结合当前局势推荐更适合的下一幕
+- 转场辅助链路：除了推荐哪一幕，还会给出转场起句、后续戏路，以及切幕后可直接继续开聊的自动起拍提示
 
 ## 使用方式 🛠️
