+++ src/skill_support/__init__.py
@@ -3,3 +3,18 @@
 
 """Shared helpers for prompt-first skill workflows."""
 
+from .scene_recommendations import (
+    build_scene_opening_message,
+    build_scene_recommendation_bundle,
+    normalize_scene_recommendation_context,
+    recommend_dialogue_scene_cards,
+    recommend_scene_cards_base,
+)
+
+__all__ = [
+    "build_scene_opening_message",
+    "build_scene_recommendation_bundle",
+    "normalize_scene_recommendation_context",
+    "recommend_dialogue_scene_cards",
+    "recommend_scene_cards_base",
+]
