+++ tests/test_packaging_docs.py
@@ -20,6 +20,8 @@ class PackagingDocsTests(unittest.TestCase):
         self.assertIn("tools/build_prompt_payload.py", manifest_text)
         self.assertIn("tools/materialize_persona_bundle.py", manifest_text)
         self.assertIn("tools/export_relation_graph.py", manifest_text)
+        self.assertIn("tools/build_scene_recommendation_payload.py", manifest_text)
+        self.assertIn("examples/scene_recommendation_context.example.json", manifest_text)
         self.assertIn("tools/verify_host_workflow.py", manifest_text)
         self.assertNotIn("runtime/zaomeng_cli.py", manifest_text)
         self.assertNotIn("runtime/src", manifest_text)
@@ -52,6 +54,8 @@ class PackagingDocsTests(unittest.TestCase):
         self.assertIn("tools/prepare_novel_excerpt.py", skill_readme)
         self.assertIn("tools/build_prompt_payload.py", skill_readme)
         self.assertIn("tools/export_relation_graph.py", skill_readme)
+        self.assertIn("tools/build_scene_recommendation_payload.py", skill_readme)
+        self.assertIn("examples/scene_recommendation_context.example.json", skill_readme)
         self.assertIn("requirements.txt", skill_readme)
         self.assertIn("references/chat_contract.md", skill_readme)
         self.assertIn("references/capability_index.md", skill_readme)
@@ -59,6 +63,8 @@ class PackagingDocsTests(unittest.TestCase):
         self.assertIn("tools/prepare_novel_excerpt.py", skill_readme_en)
         self.assertIn("tools/build_prompt_payload.py", skill_readme_en)
         self.assertIn("tools/export_relation_graph.py", skill_readme_en)
+        self.assertIn("tools/build_scene_recommendation_payload.py", skill_readme_en)
+        self.assertIn("examples/scene_recommendation_context.example.json", skill_readme_en)
         self.assertIn("requirements.txt", skill_readme_en)
         self.assertIn("references/chat_contract.md", skill_readme_en)
         self.assertIn("references/capability_index.md", skill_readme_en)
@@ -99,6 +105,7 @@ class PackagingDocsTests(unittest.TestCase):
         self.assertIn("act", clawhub_skill)
         self.assertIn("insert", clawhub_skill)
         self.assertIn("observe", clawhub_skill)
+        self.assertIn("scene_recommendation", clawhub_skill)
         self.assertNotIn("src.cli.app chat", clawhub_skill)
         self.assertNotIn("src.cli.app chat", skill_readme)
         self.assertNotIn("src.cli.app chat", skill_readme_en)
@@ -115,6 +122,8 @@ class PackagingDocsTests(unittest.TestCase):
         self.assertIn("act", contract_text)
         self.assertIn("insert", contract_text)
         self.assertIn("observe", contract_text)
+        self.assertIn("build_scene_recommendation_payload.py", contract_text)
+        self.assertIn("scene_recommendation_context.example.json", contract_text)
         self.assertIn("PROFILE.md", contract_text)
         self.assertIn("MEMORY.md", contract_text)
         self.assertIn("run_manifest.json", contract_text)
@@ -128,6 +137,8 @@ class PackagingDocsTests(unittest.TestCase):
         self.assertIn("export_graph", capability_text)
         self.assertIn("verify_workflow", capability_text)
         self.assertIn("Dialogue Stage", capability_text)
+        self.assertIn("build_scene_recommendation_payload.py", capability_text)
+        self.assertIn("scene_recommendation_context.example.json", capability_text)
         self.assertIn("references/chat_contract.md", capability_text)
         self.assertIn("examples/host_workflow_example.md", capability_text)
         self.assertNotIn("src.cli.app chat", capability_text)
@@ -141,6 +152,7 @@ class PackagingDocsTests(unittest.TestCase):
         self.assertIn("tools/verify_host_workflow.py", workflow_text)
         self.assertIn("Hand Off To Dialogue", workflow_text)
         self.assertIn("run_manifest.json", workflow_text)
+        self.assertIn("scene_recommendation_context.example.json", workflow_text)
         self.assertNotIn("src.cli.app chat", workflow_text)
 
     def test_distillation_docs_require_multi_character_differentiation(self):
