+++ zaomeng-skill/tools/_skill_support/__init__.py
@@ -12,6 +12,13 @@ from .dialogue_payloads import (
     normalize_dialogue_suggestion_context,
     parse_dialogue_suggestion,
 )
+from .scene_recommendations import (
+    build_scene_opening_message,
+    build_scene_recommendation_bundle,
+    normalize_scene_recommendation_context,
+    recommend_dialogue_scene_cards,
+    recommend_scene_cards_base,
+)
 from .persona_review import (
     PERSONA_AUTOFILLABLE_FIELDS,
     PERSONA_REVIEW_ADVANCED_GROUPS,
@@ -77,17 +84,22 @@ __all__ = [
     "build_random_self_card_messages",
     "build_dialogue_suggestion_llm_messages",
     "build_dialogue_suggestion_payload",
+    "build_scene_opening_message",
+    "build_scene_recommendation_bundle",
     "compact_dialogue_suggestion_payload",
     "delete_self_card_payload",
     "list_self_cards_payload",
     "load_persona_review_payload",
     "load_self_card_payload",
     "normalize_dialogue_suggestion_context",
+    "normalize_scene_recommendation_context",
     "normalize_self_card_fields",
     "parse_dialogue_suggestion",
     "parse_persona_field_completion_response",
     "parse_random_self_card_response",
     "read_persona_review_fields",
+    "recommend_dialogue_scene_cards",
+    "recommend_scene_cards_base",
     "resolve_persona_review_source",
     "save_persona_review_profile",
     "save_self_card_payload",
